🐛 fix(Ansible): change hosts from 'all' to '*' in multiple playbooks The hosts field in multiple Ansible playbooks was changed from 'all' to '*' to ensure that the playbook runs on all hosts. This change was made to improve consistency and avoid potential issues with running the playbook on a subset of hosts.
🚀 feat(Makefile): add make targets to change Ansible playbook hosts
Added two new make targets to change the Ansible playbook hosts to either localhost or wildcard. This will allow for more flexibility when running the playbook on different environments.

🚀 chore(Makefile): add Ansible inventory file path to all Ansible playbooks
This commit adds the Ansible inventory file path to all Ansible playbooks in the Makefile. This ensures that the Ansible playbooks are executed on the correct hosts specified in the inventory file. Additionally, two new targets are added to the Makefile to change the hosts file to either localhost or wildcard.

🐛 fix(Docker.sh): correct typo in error message
The error message had a typo in the word "specified". It was spelled as "sphecified". This commit fixes the typo.

🚀 chore(Docker.sh): update ansible playbook paths
The ansible playbook paths have been updated to reflect the new directory structure of the project. The paths have been updated to include the Launchpad directory.

🐛 fix(DockerContainerSetup.sh): fix path to Docker Compose files
The path to the Docker Compose files was incorrect, causing the containers to not start properly. The correct path has been added to the script.

🚚 chore(DockerContainerSetup.sh): update directory location for Docker Compose files
The directory location for Docker Compose files has been updated to /$directory_location/Launchpad. This change was made to improve the organization of the project and to make it easier to locate the Docker Compose files.

🚀 feat(TailscaleHostname.sh): change directory location to Launchpad and add change_hosts_to_localhost.sh script
The TailscaleHostname.sh script now changes the directory location to Launchpad instead of Docker/Containers/Tailscale. The change_hosts_to_localhost.sh script has been added to change the playbook hosts to localhost. This improves the setup process by making it more modular and flexible.

🔨 refactor(change_hosts_to_wildcard.sh): remove unnecessary comments and add input validation
✨ feat(change_hosts_to_wildcard.sh): add input prompt and validation for directory location
The script now prompts the user for the directory location of the Launchpad GitHub repository and validates that the directory exists. The script also removes unnecessary comments.