🚀 feat(Makefile): change change_hosts_to_wildcard to change_hosts_to_all The make command `change_hosts_to_wildcard` has been renamed to `change_hosts_to_all` to better reflect the change that is being made. The corresponding shell script has also been updated to replace 'hosts: localhost' with 'hosts: all' in all the playbook files and remove the 'connection: local' line after 'hosts: all' in all the playbook files.
🐛 fix(change_hosts_to_localhost.sh): change 'hosts: all' to 'hosts: localhost' in all playbook files and add 'connection: local' after 'hosts: localhost'
The script now replaces 'hosts: all' with 'hosts: localhost' in all playbook files. This change ensures that the playbook runs only on the local machine. Additionally, the script adds 'connection: local' after 'hosts: localhost' in all playbook files. This change ensures that the playbook runs locally and not over SSH.