🔥 chore(docker): remove Dashy README.md file ✨ feat(docker): add docker-compose.yml files for Double Commander, DupeGuru, Floatplane Downloader, Grocy, Joplin, and Memcached The Dashy README.md file was removed. Docker-compose files were added for Double Commander, DupeGuru, Floatplane Downloader, Grocy, Joplin, and Memcached. These files define the services, images, container names, hostnames, environment variables, volumes, ports, restart policies, labels, and networks for each service. These changes improve the maintainability and scalability of the Docker containers.
🎉 feat(Docker): add mongodb and monitorss services to docker-compose.yml
The docker-compose.yml file was created to add the mongodb and monitorss services to the Docker environment. The mongodb service is used to store data for the monitorss service. The monitorss service is a bot and web application that provides RSS feed monitoring and notification functionality. The bot and web services are configured with environment variables that are set in the docker-compose.yml file. The bot service is configured to run on the first container startup using the binfmt-starter.sh script.