✨ feat(docker-compose.yml): add new docker-compose files for cloudflared, dashy, monitorss, nginx_proxy_manager, and node-red These new docker-compose files have been added to the project to enable the deployment of cloudflared, dashy, monitorss, nginx_proxy_manager, and node-red. The files contain the necessary configurations for each service to run in a containerized environment.
🎉 feat(docker-compose.yml): add docker-compose.yml file for Photoprism
This commit adds a docker-compose.yml file for Photoprism, which is a self-hosted photo management application. The file contains the necessary configuration for running Photoprism in a Docker container, including environment variables, volumes, and network settings. The file also includes a MariaDB container for use as a database server.

🎉 feat(pterodactyl-panel): add Docker Compose files for Pterodactyl Panel
The Docker Compose files for Pterodactyl Panel have been added to the repository. The files include a README.md file with a brief description of the directory and a docker-compose.yml file that defines the Pterodactyl Panel service. The docker-compose.yml file includes the necessary configuration to run the Pterodactyl Panel service, including the image, container name, hostname, ports, volumes, environment variables, and network configuration.