🗑️ chore(traefik): remove unused README.md file The README.md file in the Traefik directory was not being used and has been removed to declutter the repository.
🎉 feat(docker-compose.yml): add docker-compose file for Photoprism 2
This commit adds a new docker-compose.yml file for Photoprism 2, which is a self-hosted photo management application. The file contains the configuration for the Photoprism 2 service and the recommended MariaDB database server. The configuration includes environment variables for the Photoprism 2 service, such as the admin user credentials, site URL, and database connection details. The volumes are also configured to mount the necessary directories for the service to function properly.

🎉 feat(docker-compose.yml): add docker-compose file for Photoprism version 3
This commit adds a new docker-compose file for Photoprism version 3. It includes the configuration for the Photoprism service and the recommended MariaDB database server. The file also includes comments with useful information and links to setup guides and troubleshooting checklists.

🚀 feat(traefik/docker-compose.yml): add traefik and dockersocket services to docker-compose file
🚀 feat(authelia): add configuration.yml and users_database.yml files to Static/Docker/authelia directory
The traefik and dockersocket services are added to the docker-compose file to enable reverse proxy functionality. The configuration.yml and users_database.yml files are added to the Static/Docker/authelia directory to enable authentication and authorization functionality. The configuration.yml file contains the configuration settings for Authelia, while the users_database.yml file contains the user information for authentication.

🆕 feat(config.yaml): add Cloudflare tunnel configuration file
Added a new configuration file for Cloudflare tunnel. The file contains the configuration for the tunnel, credentials file path, and ingress rules. The ingress rules are commented out, but there are three different options for forwarding traffic to a reverse proxy with SSL, forwarding traffic to a reverse proxy with SSL and no TLS verification, and forwarding traffic to a reverse proxy over HTTP.