🚀 chore(Makefile): add scripts for environment variable management The Makefile now includes scripts for managing environment variables. The 'all-env-variables' script extracts the names of environment variables from .env files and stores them in env_var_names.txt. The 'hashicorp-vault-pull-secrets' script authenticates to Hashicorp Vault, creates .env files from .env.example files, and retrieves secret values from Vault for environment variables in the .env files. The 'hashicorp-vault-push-secrets' script sets up authentication to Hashicorp Vault, copies .env.example files to .env files, and pushes environment variables to Vault. The 'change_hosts_to_localhost' script updates the playbook files in the Launchpad repository to use 'localhost' as the host and adds 'connection: local' to each playbook. The 'change_hosts_to_all' script reverts the playbook files in the Launchpad repository to use 'all
🆕 chore(auto-push-env.sh): add script to push environmental variables to Hashicorp Vault
This script sets the VAULT_ADDR variable to a specific URL, prompts the user for their username and password, and uses the docker command to authenticate to Hashicorp Vault using the client token authentication method and the provided username and password. It then finds all files with the .env.example extension in the current directory and its subdirectories and copies each file to a new file with the same name but with a .env extension, if a file with that name does not already exist. It then finds all files with the .env extension in the current directory and its subdirectories and reads their contents into an array called env_vars, filtering out lines that start with a # (comments) or are empty. It loops through the array env_vars and for each element, it splits the element into a name and value field using the = character as a delimiter. It assigns the name field