chore(dependencies): Update dependencies.sh script chore(setup): Remove update_permissions function from setup.sh script
Update the dependencies.sh script to add the ~/.local/bin directory to the PATH for the current session and all future sessions. This ensures that the Ansible installation is operational and can be accessed properly. Also, add a check to avoid duplicating the PATH entry in ~/.bashrc.

Remove the update_permissions function from the setup.sh script as it is no longer needed.