chore(vault): Update vault version and environment variables chore(PCSMenu): Update git clone command in PCSInstall.sh
Update the version of the vault image in the docker-compose.yml file from 1.13.3 to 1.15. Also, update the environment variables in the .env.example file to use 'http://0.0.0.0:8200' as the vault address.

In the PCSInstall.sh script, update the git clone command to remove the branch specification and clone the repository without specifying a branch.