fix(vscode): update extensions and settings feat(Docker): add Dockerfile and update docker-compose.yml refactor(Scripts/Vault): refactor vault-pull.go and vault-push.go
Update the extensions.json file to include the "golang.go" and "ms-azuretools.vscode-docker" extensions. Update the settings.json file to set the default formatter for Go and Dockerfile files.

Add a new Dockerfile in the "Docker/Both/codeserver" directory to build a custom image for code-server. Update the docker-compose.yml file to use the new Dockerfile.

Refactor the vault-pull.go and vault-push.go scripts in the "Scripts/Vault" directory to improve code readability and remove unnecessary code duplication.