feat(ansible): add playbook for user creation feat(ansible): add playbook for zsh powerlevel10k configuration fix(scripts): remove comment in main.go
Added a new playbook 'user-creation.yml' to create a new user on all hosts. This playbook prompts for the username and password of the new user and adds the user to the sudo group.

Added a new playbook 'zsh_powerlevel10k.yml' to configure the zsh shell with the Powerlevel10k theme. This playbook clones the Powerlevel10k theme and zsh-syntax-highlighting plugin into the Oh-My-Zsh custom themes and plugins folders for each user, excluding root. It also checks if the zsh-history-substring-search plugin exists and clones it if it doesn't.

In the 'main.go' script, removed a commented line that was not needed.