fix(PCSMenu): update path to PersonalizationFunc.sh fix(PCSMenu): update path to PCSFunc.sh in check_update function feat(PCSMenu): display OS information in menu_bar function
The path to the PersonalizationFunc.sh file is updated to /opt/Wolflith/PCSRustMenu/src/PersonalizationFunc.sh to fix a bug. The path to the PCSFunc.sh file in the check_update function is also updated to /opt/Wolflith/PCSRustMenu/src/PCSFunc.sh. Additionally, the menu_bar function is updated to display the OS information.